Explicitly specify int32/int64 (first installment) (#10)
* Specify 32-bit/64-bit ints explicitly

* more: psdes.h/cpp

* more

+++ src/random/int_random.h
@@ -8,7 +8,7 @@
 #define INT_RANDOM_H
 
 unsigned get_random_unsigned(void);
-unsigned long long get_random_ull(void);
+uint64_t get_random_ull(void);
 int get_random_int(void);
 
 #endif // INT_RANDOM_H
